Python/requirements.txt
zeroconf
cryptography
argon2-cffi